[NFC] Rename some PALs to make the naming more consistent.
