+++ src/pal/pal.h
@@ -10,7 +10,7 @@ namespace snmalloc
 // If simultating OE, then we need the underlying platform
 #if !defined(OPEN_ENCLAVE) || defined(OPEN_ENCLAVE_SIMULATION)
 #  include "pal_apple.h"
-#  include "pal_free_bsd_kernel.h"
+#  include "pal_freebsd_kernel.h"
 #  include "pal_freebsd.h"
 #  include "pal_linux.h"
 #  include "pal_netbsd.h"
@@ -33,11 +33,11 @@ namespace snmalloc
 #  elif defined(FreeBSD_KERNEL)
     PALFreeBSDKernel;
 #  elif defined(__FreeBSD__)
-    PALFBSD;
+    PALFreeBSD;
 #  elif defined(__NetBSD__)
     PALNetBSD;
 #  elif defined(__OpenBSD__)
-    PALOBSD;
+    PALOpenBSD;
 #  else
 #    error Unsupported platform
 #  endif
